updated fish and bootstrap config

+++ environments/linux/scripts/commons/config.sh
@@ -14,11 +14,13 @@ if [ ! -d ~/.fonts ]; then
   mkdir -p ~/.fonts
 fi
 pushd ~/.fonts
-curl https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Monofur.zip -O -J -L
-curl https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Hack.zip -O -J -L
-unzip Monofur.zip -d Monofur
-unzip Hack.zip -d Hack
-fc-cache -fv ~/.fonts
-rm ~/.fonts/Monofur.zip
-rm ~/.fonts/Hack.zip
+if [ ! -d Monofur ]; then
+  curl https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Monofur.zip -O -J -L
+  curl https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/Hack.zip -O -J -L
+  unzip Monofur.zip -d Monofur
+  unzip Hack.zip -d Hack
+  fc-cache -fv ~/.fonts
+  rm ~/.fonts/Monofur.zip
+  rm ~/.fonts/Hack.zip
+fi
 popd
